Just moved character, and implementations into other files
